extracted program logic into Main
