+++ src/s1/t1/n1/exercici2/Cotxe.java
@@ -1,13 +1,5 @@
 package s1.t1.n1.exercici2;
 
-public class CotxeTest {
-    void main(String[] args) {
-        Cotxe car = new Cotxe("Camaro", 300); 
-        car.accelerar();
-        Cotxe.frenar();
-    }
-        
-}
 
 class Cotxe {
     static final String MARCA = "Ford";
