+++ src/s1/t1/n1/exercici2/Main.java
@@ -1,0 +1,9 @@
+package s1.t1.n1.exercici2;
+
+public class Main {
+    void main (String[] args){
+        Cotxe car = new Cotxe("Camaro", 300); 
+        Cotxe.frenar();
+        car.accelerar();
+    }
+}
